Allow "ad" tool suite to operate with CNID backends other then "dbd"
Originally this check was to ensure that one didn't try running one of the "ad" tools against a volume using the `last` CNID backend. The current code assumes that the only compatible backend is `dbd` and was never updated when the MySQL CNID was added to Netatalk.

Instead of checking specifically for the `dbd` backend, check if the volume is running the `last` backend and error out. This allows all other present and future CNID options to work with the tools.

Note: The MySQL CNID has been tested with the `ad` tools and is working properly.